+++ src/utilities/getTodaysFormattedDate.js
@@ -1,5 +1,5 @@
 import { startOfToday, formatISO9075 } from "date-fns";
 
-export const todaysDate = formatISO9075(new Date(startOfToday()), {
+export const getTodaysFormattedDate = formatISO9075(new Date(startOfToday()), {
   representation: "date",
 });